% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostPM-package.R
\docType{package}
\name{boostPM-package}
\alias{boostPM-package}
\alias{boostPM}
\title{boostPM: Unsupervised Tree Boosting for Learning Probability Distributions}
\description{
Implements the unsupervised tree boosting method of Awaya and Ma (2024) for
fitting tree-ensemble probability distributions, evaluating fitted
densities, and generating samples.
}
\references{
Awaya, N. and Ma, L. (2024). Unsupervised Tree Boosting for Learning
Probability Distributions. \emph{Journal of Machine Learning Research}, 25,
1--52.
}
\seealso{
\code{\link[=fit_boostpm]{fit_boostpm()}}, \code{\link[=predict.boostPM_fit]{predict.boostPM_fit()}},
\code{\link[=simulate.boostPM_fit]{simulate.boostPM_fit()}}, and \code{\link[=plot.boostPM_fit]{plot.boostPM_fit()}}.
}
\author{
\strong{Maintainer}: Naoki Awaya \email{nawaya@waseda.jp}

}
