% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-methods.R
\name{print.boostPM_fit}
\alias{print.boostPM_fit}
\title{Print a Fitted boostPM Distribution}
\usage{
\method{print}{boostPM_fit}(x, ...)
}
\arguments{
\item{x}{A fitted object returned by \code{\link[=fit_boostpm]{fit_boostpm()}}.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Prints a compact description of a boostPM_fit object.
}
\examples{
set.seed(10)
x1 <- stats::rbeta(60, shape1 = 2, shape2 = 5)
x <- cbind(x1 = x1, x2 = stats::rbeta(60, 2 + 4 * x1, 3))
fit <- fit_boostpm(
  x,
  Omega = cbind(lower = c(0, 0), upper = c(1, 1)),
  add_noise = FALSE,
  ntree_max_marginal = 1,
  ntree_max_dependence = 1,
  max_resol = 1
)
print(fit)

}
